Send update Actor available

+++ routes/api.php
@@ -35,6 +35,9 @@ Route::prefix('GET')->group(function () {
 
     //DEMANDE DE SUPPRESSION
     Route::get('delete/demand/{id}', 'ActorController@sendDelete')->name('actor.sendDelete');
+
+    //DEMANDE DE MODIFICATION
+    Route::get('update/demand/{id}', 'ActorController@sendUpdate')->name('actor.sendUpdate');
 });
 
 //!ROUTES EN POST
@@ -71,13 +74,16 @@ Route::prefix('admin')->group(function () {
     Route::prefix('PUT')->group(function () {
         //MODIFIER BUFFER
         Route::put('buffer/{id}', 'BufferController@update')->name('buffer.update');
+
+        //MODIFIER ACTORS
+        Route::put('actor/{id}', 'ActorController@update')->name('actor.update');
     });
 
     //?ROUTES EN DELETE
     Route::prefix('DELETE')->group(function () {
 
         // SUPPRIMER ACTOR
-        Route::delete('actor/{id}', 'ActorController@destroy')->name('actor.destroy');
+        Route::delete('actor/{id}', 'ActorController@destroy')->name('actor.destroy')->middleware('auth:admin');
 
         // SUPPRIMER BUFFER
         Route::delete('buffer/{id}', 'BufferController@destroy')->name('buffer.destroy');
